Reimplemented the energy tracking for identities (#115)
* Refactoring some stuff for energy

* Fix an issue with i128 query params

* Infinite budget in Standalone

* Energy and crash fixes

* Hopefully fixed the test that now has energy

* Addresses Centril's comments

* Cargo fmt

---------

+++ crates/core/src/messages/control_db.rs
@@ -13,8 +13,11 @@ pub struct IdentityEmail {
 #[derive(Clone, PartialEq, Serialize, Deserialize)]
 pub struct EnergyBalance {
     pub identity: Identity,
-    /// How much budget is remaining for this identity.
-    pub balance: i64,
+    /// The balance for this identity this identity.
+    /// NOTE: This is a signed integer, because it is possible
+    /// for a user's balance to go negative. This is allowable
+    /// for reasons of eventual consistency motivated by performance.
+    pub balance: i128,
 }
 
 #[derive(Clone, PartialEq, Serialize, Deserialize)]
